refactor: add private folder
