+++ src/app/info/_components/purchase-card.tsx
